# backend/swagger-docs.json
{"openapi":"3.0.0","info":{"title":"Fake Ecommerce Marketplace Api","description":"Fake Ecommerce Marketplace Api","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3333/"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"PaginationInfos":{"type":"object","properties":{"total":{"type":"number"},"skip":{"type":"number"},"limit":{"type":"number"}},"required":["total","skip","limit"]},"ObjectID":{"type":"object","properties":{}},"Product":{"type":"object","properties":{"_id":{"$ref":"#/components/schemas/ObjectID"},"name":{"type":"string","maxLength":50},"price":{"type":"number"},"shortDescription":{"type":"string","maxLength":120},"imageUrl":{"type":"string"},"createdAt":{"type":"number"}},"required":["_id","name","price","createdAt"]},"ListProducts":{"type":"object","properties":{"pagination":{"$ref":"#/components/schemas/PaginationInfos"},"results":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}},"required":["pagination","results"]},"CreateProductDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":120},"price":{"type":"number"},"shortDescription":{"type":"string","maxLength":120},"imageUrl":{"type":"string"}},"required":["name","price","imageUrl"]},"UpdateProductDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":120},"price":{"type":"number"},"shortDescription":{"type":"string","maxLength":120},"imageUrl":{"type":"string"}},"required":["name","price","imageUrl"]},"PatchProductDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":120},"price":{"type":"number"},"shortDescription":{"type":"string","maxLength":120},"imageUrl":{"type":"string"}}}}},"paths":{"/v1/products":{"get":{"operationId":"ProductsController_getMany","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"skip","required":false,"in":"query","schema":{"default":0,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"sort-createdAt","required":false,"in":"query","schema":{"enum":["1","-1",1,-1],"type":"string"}}],"responses":{"200":{"description":"Productes obtido(a)s com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListProducts"}}}}},"tags":["products"]},"post":{"operationId":"ProductsController_store","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDTO"}}}},"responses":{"201":{"description":"Product cadastrado(a) com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["products"]}},"/v1/products/{id}":{"get":{"operationId":"ProductsController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Product obtido(a) com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["products"]},"put":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDTO"}}}},"responses":{"200":{"description":"Product atualizado(a) com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["products"]},"patch":{"operationId":"ProductsController_patch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchProductDTO"}}}},"responses":{"200":{"description":"Product atualizado(a) com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["products"]},"delete":{"operationId":"ProductsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Product deletado(a) com sucesso."}},"tags":["products"]}}}}